JVMCBC-1161: Cluster disconnect is intermittently timing out
Shutdown currently causes a number of updated configs
to be pushed.  Core.shutdown closes all buckets, each of which emits
an updated config with that bucket removed.
Then DefaultConfigurationProvider calls
closeBucketIgnoreShutdown once per bucket,
which (again) emits an updated config with that
bucket removed from bucketConfigs.  Then
doOnTerminate emits a final config that also has globalConfig
removed.

These multiple configs are resulting in at least one race condition,
where maybeRemoveNode refuses to remove a node as it still has
a bucket config, resulting in shutdown timing out as it's waiting
for nodes to be removed that won't be.

So part of this changeset is to have DefaultConfigurationProvider
emit just one final empty config during shutdown.

However, that wasn't enough to eliminate races.  If a regular
global or bucket config came in just before shutdown, it
would continue to race with that empty config, resulting in
Core.shutdown once again waiting for nodes to be removed that
never would be.

An attempt to fix this by having
DefaultConfigurationProvider.pushConfig() skip the config if it
was shutting down helped reduce the window substantially, but
did not close it completely.

So the second part of the changeset, is to have Core.shutdown
itself, emit one empty config after DefaultConfigurationProvider
has completed.  At this point it shouldn't be possible for any
configs (including those that arrived just before shutdown) to
be emitted, so it shouldn't be possible for this to race anymore.

This second fix possibly renders the first fix redundant.  However,
it seems anyway cleaner to have DefaultConfigurationProvider only emit
a single config during shutdown, and the fixes have been thoroughly
tested together now.  So I've left both in place.

All changes only take effect during shutdown.

Change-Id: If840ce2a19e07bddce4724a26bce5014f6ea3971