JVMCBC-1185 Move KV CRUD operations into core
Motivation
----------
Adding a CoreKvOps interface paves the way for slotting
in alternative protocols. This is groundwork for the
Protostellar protocol.

Pushing functionality down into core-io means less work
for the JVM language wrappers (currently Scala & Kotlin).

Modifications
-------------
Add a CoreKvOps interface, and a "Classic" implementation
that uses the existing Memcached protocol code.

CoreKvOps currently covers the basic KV CRUD operations.
Subdocument and binary operations are still todo.

Add a repackaged version of gRPC to core-io, along with
prototype stubs for the Protostellar protocol.

Migrate the test code in AsyncScopeTest to
KeyValueCollectionIntegrationTest, because
mocking is a pain.

Change-Id: I54fe6ab472398b8adcdd0cf1d83b6a25326c13fd