Deshadowing WIP
This is a very rough draft on what a 'deshadowed' core-io - one
that does not shadow Netty and other dependencies - could look like.

To do this change for real:
We will need to automate the package renaming as part of the release
process, and publish a separate core-io that only Quarkus will use.

Change-Id: I07b90d63f1c4f63ea360b37c05a31bf989edf0b2